refactors factory to create objects for entire rail system

+++ tests/map/test_metro_north.cpp
@@ -16,7 +16,7 @@ TEST_F(MetroNorthGraphTest, CreatesNodeStationsSuccessfully)
 {
     using namespace Transit::Map;
 
-    int n{10};
+    size_t n{10};
     std::vector<int> expected_ids{};
 
     try
